+++ test/adversarialLearningAgentTest.cpp
@@ -257,10 +257,9 @@ TEST_F(adversarialLearningAgentTest, EvalAllRootsSequential)
 
     auto le2 = FakeClassificationLearningEnvironment();
     Learn::AdversarialLearningAgent laNotCopyabe(le2, set, params);
-
-    ASSERT_THROW(
-        laNotCopyabe.evaluateAllRoots(0, Learn::LearningMode::TRAINING),
-        std::runtime_error);
+    // not copyable + sequential => should work
+    ASSERT_NO_THROW(
+        laNotCopyabe.evaluateAllRoots(0, Learn::LearningMode::TRAINING));
 }
 
 TEST_F(adversarialLearningAgentTest, EvalAllRootsParallel)
@@ -283,6 +282,13 @@ TEST_F(adversarialLearningAgentTest, EvalAllRootsParallel)
     ASSERT_EQ(result.size(), la.getTPGGraph().getNbRootVertices())
         << "Number of evaluated roots is under the number of roots from the "
            "TPGGraph.";
+
+    auto le2 = FakeClassificationLearningEnvironment();
+    Learn::AdversarialLearningAgent laNotCopyabe(le2, set, params);
+    // not copyable + multithread => exception
+    ASSERT_THROW(
+        laNotCopyabe.evaluateAllRoots(0, Learn::LearningMode::TRAINING),
+        std::runtime_error);
 }
 
 TEST_F(adversarialLearningAgentTest, EvalAllRootsGoodResults)